==> data-base-constitution/1.Fishing.Rmd <==
---
title: "Fishing from seed"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidytext)
library(dplyr)
library(stopwords)
library(ggplot2)
library(lubridate)
library(stringr) # str_detect
library(rtweet)
library(Dict)
readRenviron("~/.Renviron")

auth <- rtweet_app()
# access_token <- "AAAAAAAAAAAAAAAAAAAAAIcKLwEAAAAA9Ve0dVTIyaecdx0XwYfOpoLYQek%3D9g08wPfTo0IQSPu16Ysj0aYpXkr11ETwzN9ZFhNcvCx3lC5RQl"
```

We start from a list of journalists from different newspapers.

```{r}
journalists.seed <- read.csv("../data/journalistes.csv", sep = ";")
```

```{r}
# get someone's friends (followers IDs)
fishing.probable.journalists <- data.frame("user" = c(),
                                           "user_id" = c())

count = 0

# because of limits, it is not possible to fish everyone from the table directly. 
# indeed, apparently , retryonratelimit = TRUE was not working so well
for (journalist in journalists.seed$pseudo[13:71]){
  
  if (count < 4){
    
    count = count + 1
    
    print(paste0("doing ", journalist))
    fds <- get_friends(journalist, retryonratelimit = TRUE)
    fishing.probable.journalists <- rbind(fishing.probable.journalists, fds)
    
    
  } else {
    
    date_time <- Sys.time()
    print(paste0("starting the break at ", date_time))
    
    # start the nap!
    Sys.sleep(16*60)  
    
    date_time <- Sys.time()
    print(paste0("ending the break at ", date_time))
    
    print(paste0("doing ", journalist))
    fds <- get_friends(journalist, retryonratelimit = TRUE)
    fishing.probable.journalists <- rbind(fishing.probable.journalists, fds)
    
    # start again
    count = 1
  }
}

```

```{r}
# 108078 contacts
fishing.probable.journalists <- distinct(fishing.probable.journalists)
```

```{r}
# store data
save(fishing.probable.journalists, file = "./fishing2.RData")
```
