==> data-base-constitution/3.ScrapTimeline.Rmd <==
---
title: "Constitute data base of journalists"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidytext)
library(dplyr)
library(tidyr)
library(stopwords)
library(ggplot2)
library(stringi)
library(Dict)
library(Xplortext)
library(DescTools)
library(stringr)
library(FactoMineR)

source("./param.R")
```

```{r}
library(rtweet)
auth <- rtweet_app()
# access_token <- "AAAAAAAAAAAAAAAAAAAAAIcKLwEAAAAA9Ve0dVTIyaecdx0XwYfOpoLYQek%3D9g08wPfTo0IQSPu16Ysj0aYpXkr11ETwzN9ZFhNcvCx3lC5RQl"
```

Load the previous batches.

```{r}
load("../data/journalist-and-description.RData")
total <- probable.journalist
rm(probable.journalist)
```

Take the subset of journalist for which we have at least one redaction.

```{r}
has.redactions <- rep(0, nrow(total))

for (redac in redactions$keys){
  has.redactions <- has.redactions + total[, redac]
}

total$has.redactions <- has.redactions

hist(total$has.redactions)
```

```{r}
total <- total[total$has.redactions > 0,]
```


```{r}
# initialisation
timeline <- get_timeline("lemondefr", n = 1, lang = "fr", include_rts = FALSE)
timeline$screen_name <- rep("LeMonde", nrow(timeline))
```


```{r}
count = 0
store = 10 
# filling the table
for (name in total$screen_name[2500:3000]){
  
  if (count < 501){
    
      count = count + 1
      print(paste0("doing ", name))
      try(new_row <- get_timeline(name, n = 5000, lang = "fr", include_rts = FALSE), silent = TRUE)
      new_row$screen_name <- rep(name, nrow(new_row))
      timeline <- rbind(timeline, new_row)
  
  } else {
      
      date_time <- Sys.time()
      print(paste0("starting the break at ", date_time))
      
      new.name <- paste0("../data/timeline", store, ".Rdata")
      save(timeline, file = new.name)
      store = store + 1
      timeline <- timeline[1,]
      
      # start the nap!
      Sys.sleep(16*60)  
      date_time <- Sys.time()
      print(paste0("ending the break at ", date_time))
      
      count = 0
  }
}
```


```{r}
save(timeline, file = "../data/officials.RData") 
```

```{r}
officials <- Dict$new(
  LeMonde = c("lemondefr", "lemonde_pol"),
  LeFig = c("Le_Figaro", "FigaroVox", "Figaro_culture"),
  Libe = c("libe"),
  Mediapart = c("Mediapart", "MediapartBlogs"),
  Point = c("LePoint"),
  Valeur = c("Valeurs", "VAplus_"),
  Obs = c("lobs"),
  Marianne = c("Marianne_TVoff", "MarianneleMag"),
  AFP = c("afpfr"),
  Blast = c("blast_france"),
  Cnews = c("CNEWS"),
  Express = c("LEXPRESS"),
  Inter = c("franceinter", "franceinfo"),
  Europe = c("Europe1"),
  BFM = c("BFMTV"),
  Lacroix = c("LaCroix"),
  Parisien = c("le_Parisien"),
  LaProvence = c("laprovence"),
  OuestFrance = c("OuestFrance"),
  VoixDuNord = c("lavoixdunord"),
  LeProgres = c("Le_Progres"),
  SudOuest = c("SO_Paysbasque"),
  DauphineLibere = c("ledauphine"),
  EstRep = c("lestrepublicain"))
```


```{r}
for (journal in officials$keys){
  for (id in officials[journal]){
    print(paste0("doing ", id))
    try(new_row <- get_timeline(id, n = 10000, lang = "fr", include_rts = FALSE), silent = TRUE)
    new_row$screen_name <- rep(journal, nrow(new_row))
    timeline <- rbind(timeline, new_row)
  }
}
```
